Not the most useful, but cool update.
We now have support for css keyframe animation of the devices icon.
Currently the Lifx icon will kinda follow the lights. A few caveat
    - The transition path is not the same between CSS and Lifx, this
      means the the start and end colours are the same, but intermediate
      colours are not.

    - The end colour of a Lifx pulse command is sometimes the starting
      colour and somtimes the end colour when asking for it to be the end
      colour. The icon colour is always the starting colour.

    - keyframe animation can be impeded by the reception of status messages

    - Whilst the name of the keyframe animation is always different, the
      class that triggers the animation must always be
            run-animation-<device name>

    - From my various attempts, it seems that 2 consecutive animations can not
      have the same name.

All this is bleeding edge, keyframe animation status is W3C Working Draft. This means it
could break soon.

In order to get an icon animated, the "icon-status" dictionary in the message "content"
must have an "animation" key containing the keyframe definition and the defintion
of the class "run-animation-<device name>". WebBuddy will then be able to animate the icon.