This is indroducing KodiBuddy to the bus. KodiBuddy enables AutoBuddy to integrate Kodi media system. With KodiBuddy you can     - Turn on a Kodi instance. For this to work       it must be configured to accept Wakeup On LAN       packets
    - Shutdown Kodi... carefull it can shutdown a computer!

    - Start Album/Song/Music Video/Movie by name, genre, random
      should probably add artist....

KodiBuddy comes with a small remote control that can be used to control
the player.

You need to allow Kodi to be controlled by programs on other computer for
this to work. KodiBuddy uses JSON-RPC over websockect for a nice, two-way
communication with the device. The downside is that there seem to be no
authentication mechanism.

KodiBuddy uses the new "info" function of WebBuddy. It will display info about
what is currently playing on the device. If some image is available it will try
to display it ... but... it will try to access the image using Kodi HTTP interface,
given this may require authentication, you will need to configure the authentication
for that device AND you will need to access the HTTP interface (and login) at least
once before the images will be displayed correctly. Since the image is accessed using
the credential in the URL this is NOT safe and is apparently scheduled to be disabled
in Chrome and Firefox. This bit calls for a Kodi Add-on that would expose those images
over HTTP without the need to authenticate.

ScannerBuddy now accept a "scan request" command with a specific service name
to trigger a specific scan. This is used by KodiBuddy to initiate a scan for
a known Kodi instance that has not been seen by ScannerBuddy yet after a power
on command has been issued to the device. it is also useful in case the device
IP address changes.

buddylib  simple cleanup of the device types

buddyguilib  The "rteffect" attribute now works like this
            undefined or = 0 The XML will not be rendered in case
                             of "realtime" use.

            == 1             The XML will always be rendered.

            == -1            Rendered ONLY for "realtime" used

            Note that when expanding device commands, we use a
            "Choice" control to list the possible command names
            so WebBuddy (buddy.js) now also look at the attribute
            to make sure it needs to be included. (Were it to be
            included, it would simply show nothing when selected)

WebBuddy  beside checking for "rteffect" when expanding devices' commands,
          WebBuddy now support device information. If any is available
          it can be displayed by doing
                    Ctrl Click
          on the device. KodiDevice can thus display info about what is
          playing on the device. Lifx should provide some info soon. For
          flic, it depends on Shortcut Labs.