Initial commit of I2CBuddy.     It use plugins to control various I2C sensors. At this time         bmp180 (pressure, temperature) 	bh1785 (luminosity) 	htu21d (humidity, atmospheric pressure)     are supported
    This will generate "measurement" events at a configurable
    rate.

    One need to enable the plugin for the relevant measurments to be
    broadcast

In ControlBuddy, a bridge/device can now request not to receive
events, this reduces the network traffic for those entities not processing
events. This is done by sending a message with a content_type of "mute events".
A message with a content_type "get events" can restart the sending of events.